// utils/records/types.ts
export type TPeriod = 'weekly' | 'monthly' | 'quarterly' | '';
export type TWeekDay = '1' | '2' | '3' | '4' | '5' | '6' | '0' | '';
export type THolidays = 'holidaysNoMatter' | 'beforeHoliday' | 'afterHoliday' | '';
export type TMonthDays =
  | '1'
  | '2'
  | '3'
  | '4'
  | '5'
  | '6'
  | '7'
  | '8'
  | '9'
  | '10'
  | '11'
  | '12'
  | '13'
  | '14'
  | '15'
  | '16'
  | '17'
  | '18'
  | '19'
  | '20'
  | '21'
  | '22'
  | '23'
  | '24'
  | '25'
  | '26'
  | '27'
  | '28'
  | '29'
  | '30'
  | '31'
  | '';

export type TQuarter = 'Янв, Апр, Июл, Окт' | 'Фев, Май, Авг, Ноя' | 'Мар, Июн, Сен, Дек' | '';